@@ -1,21 +1,31 @@
 import { Injectable } from '@angular/core';
-import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
+import {
+  HttpRequest,
+  HttpHandler,
+  HttpEvent,
+  HttpInterceptor
+} from '@angular/common/http';
 import { Observable, throwError } from 'rxjs';
-import { catchError, mergeMap } from 'rxjs/operators';
+import { mergeMap, catchError } from 'rxjs/operators';
 import { AuthService } from '../services/auth.service';
+import { AppConfigService } from '../services/app-config.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthInterceptor implements HttpInterceptor {
 
-  constructor(private auth: AuthService) {
+  constructor(private auth: AuthService, private config: AppConfigService) {
   }
 
   intercept(
     req: HttpRequest<any>,
     next: HttpHandler
   ): Observable<HttpEvent<any>> {
+    if (!req.url.startsWith(this.config.backendUrl)) {
+      return next.handle(req);
+    }
+
     return this.auth.getTokenSilently$().pipe(
       mergeMap(token => {
         const tokenReq = req.clone({
@@ -23,7 +33,10 @@ export class AuthInterceptor implements HttpInterceptor {
         });
         return next.handle(tokenReq);
       }),
-      catchError(err => throwError(err))
+      catchError(err => {
+        console.log('here', err);
+        return throwError(err);
+      })
     );
   }
 }
